kill subprocess only if not ready in timeout

+++ source/subprocess.ts
@@ -31,26 +31,30 @@ export class Subprocess {
   public static onStderr = (r: { stderr: Buffer, cmd: string, args: string[] }): void => undefined;
 
   public static spawn = (cmd: string, rawArgs: (string | number)[], readiness_indicator?: string, env?: { [k: string]: string }): Promise<ChildProcess> => new Promise((resolve, reject) => {
-    const ready = false;
+    let ready = false;
     const args = rawArgs.map(String);
     const p: ChildProcess = child_process.spawn(cmd, args, { env });
     PROCESSES.push(p);
     p.stdout.on('data', (stdout: Buffer) => {
       Subprocess.onStdout({ cmd, args, stdout: stdout });
       if (readiness_indicator && !ready && stdout.indexOf(readiness_indicator) !== -1) {
+        ready = true;
         resolve(p);
       }
     });
     p.stderr.on('data', (stderr: Buffer) => {
       Subprocess.onStderr({ cmd, args, stderr: stderr });
       if (readiness_indicator && !ready && stderr.indexOf(readiness_indicator) !== -1) {
+        ready = true;
         resolve(p);
       }
     });
     if (readiness_indicator) {
       setTimeout(() => {
-        reject(new ProcessNotReady(`Process did not become ready in ${SPAWN_READINESS_TIMEOUT} by outputting <${readiness_indicator}>`, [cmd].concat(rawArgs as string[])));
-        p.kill();
+        if (!ready) {
+          reject(new ProcessNotReady(`Process did not become ready in ${SPAWN_READINESS_TIMEOUT} by outputting <${readiness_indicator}>`, [cmd].concat(rawArgs as string[])));
+          p.kill();
+        }
       }, SPAWN_READINESS_TIMEOUT);
     } else {
       resolve(p);
